10-mart-efficient-code/product/pyproject.toml
[tool.poetry]
name = "mart"
version = "0.1.0"
description = ""
authors = ["Shoaib <iamshoaibxyz@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
sqlmodel = "^0.0.19"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
httpx = "^0.27.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = "^2.8.0"
aiokafka = "^0.10.0"
protobuf = "^5.27.2"
psycopg = "^3.2.0"
pytest = "^8.2.2"
psycopg2 = "^2.9.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
